Add basic support for dsk images
This format is used by the Ostro Project
(https://github.com/ostroproject/ostro-os), and its basically an raw
image (`.img`) that can be optionally flashed with `bmaptools` to
improve the writing speed.

`bmaptools` support is something we want to deal in the near future, but
at least for now, we add basic support for `.dsk` files as if they were
normal `.img` files.

See: https://github.com/resin-io/etcher/issues/491
Signed-off-by: Juan Cruz Viotti <jviottidc@gmail.com>