Separate archive functions from image archive algorithm
We support the concept of "archive images", which at the time of this
writing, represents archives containing one and only one image file in
it, which is extracted automatically.

The logic that determines if an archive is a valid "archive image" is
currently tied to its only implementation (zip). This PR separates the
actions that interact over specific archive types (called "archive
hooks"), and the actual algorithm that uses these facilities to analyze
and extract "archive images".

Signed-off-by: Juan Cruz Viotti <jviottidc@gmail.com>