Replace `.getEstimatedFinalSize` with a more generic `getImageMetatada`
Instead of having a function that just returns the estimated final size,
it makes more sense to have a function that returns every image metadata
it can get from the image, including but not limited to the estimated
final size.

When being used on Zip archives with extra metadata, this function
returns the image name, url, logo, etc.

Signed-off-by: Juan Cruz Viotti <jviottidc@gmail.com>