Remove `estimatedSize` property from metadata
We currently show the following size-related properties:

- `size`
- `estimatedUncompressedSize`
- `estimatedSize` (based on the previous two)

Which is really confusing for the end user.

As a solution, we only return `size` and `estimatedUncompressedSize`,
which have obvious meaning, and let the end user determine what a final
"estimated size" is.

The change introduced above allows for some small refactoring, which is
done on this PR as well.

Signed-off-by: Juan Cruz Viotti <jviotti@openmailbox.org>